Moving more stuff into utils and namespacing them
