+++ lib/wake/utils/errors.rb
@@ -1,0 +1,5 @@
+module Utils
+  class Error < StandardError; end
+  class CannotCreate < StandardError; end
+  class CannotUpdate < StandardError; end
+end
